Additional Pubsub integration tests
For XEP-0060 § 7.1.2. Asserts that publishing an item with the same ID
overwrites the previous item with that ID.

For XEP-0060 § 8.4.1: Assert that the server send a notification to
subscribers when deleting a node that exists.

Test for XEP-0060 § 8.4.3.2: Asserts that the server returns an
'item-not-found' error response when deleting a node that does not
exist.

Tests for non-optional ('MUST') conditions in XEP-0060 § 6.2

Tests for non-optional ('MUST') conditions in XEP-0060 § 6.1.6

Tests for non-optional ('MUST') conditions in XEP-0060 § 6.1.4

Tests for non-optional ('MUST') error conditions upon
subscription (XEP-0060 § 6.1.3).

Co-authored-by: Paul Schaub <vanitasvitae@fsfe.org>
Co-authored-by: Florian Schmaus <flo@geekplace.eu>